+++ doc/whats_new.rst
@@ -533,6 +533,8 @@ BUG
 
     - Fix :class:`mne.realtime.StimServer` by removing superfluous argument ``ip`` used while initializing the object by `Mainak Jas`_.
 
+    - Fix removal of projectors in :func:`mne.preprocessing.maxwell_filter` in ``st_only=True`` mode by `Eric Larson`_
+
 API
 ~~~